+++ src/components/feed-list/types.ts
@@ -16,3 +16,7 @@ export enum FeedListItemStatus {
     PROCESSING = 'Готовится',
     COMPLETED = 'Выполнен',
 }
+
+export type FeedListProps = {
+    data: Array<FeedListItemContent>;
+};
